requirements.txt
gunicorn==23.0.0
psycopg2-binary==2.9.10
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-dotenv==1.0.1
requests==2.32.3
werkzeug==3.1.3
pypdf>=4.0.0
